Refine rotated pill pad repro
