--- tests/repros/repro-imported-rotated-pill-pads-missing-3d/INA240A1PWR.tsx
@@ -1,34 +1,0 @@
-import type { ChipProps } from "@tscircuit/props"
-
-const pinLabels = {
-  pin1: ["NC"],
-  pin2: ["IN_POS"],
-  pin3: ["IN_NEG"],
-  pin4: ["GND"],
-  pin5: ["VS"],
-  pin6: ["REF2"],
-  pin7: ["REF1"],
-  pin8: ["OUT"],
-} as const
-
-export const INA240A1PWR = (props: ChipProps<typeof pinLabels>) => {
-  return (
-    <chip
-      pinLabels={pinLabels}
-      supplierPartNumbers={{
-        jlcpcb: ["C93965"],
-      }}
-      manufacturerPartNumber="INA240A1PWR"
-      footprint="dfn8_pillpads_p0.65mm_w7.3082mm_pw0.353mm_pl1.454mm_pin1location(leftside,bottom)"
-      cadModel={{
-        objUrl:
-          "https://modelcdn.tscircuit.com/easyeda_models/assets/C93965.obj?uuid=2d0fd2703afb4f81a9dfc54e2181a624",
-        stepUrl:
-          "https://modelcdn.tscircuit.com/easyeda_models/assets/C93965.step?uuid=2d0fd2703afb4f81a9dfc54e2181a624",
-        pcbRotationOffset: 0,
-        modelOriginPosition: { x: 0, y: 0, z: 0 },
-      }}
-      {...props}
-    />
-  )
-}
